# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для приложения карточек МозгоЁмка",
        "title": "МозгоЁмка API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/auth/forgot-password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Запросить ссылку для сброса пароля",
                "parameters": [
                    {
                        "description": "email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.ForgotPasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в систему",
                "parameters": [
                    {
                        "description": "email + password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.AuthLoginResponse"
                        }
                    },
                    "401": {
                        "description": "неверный email или пароль",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    },
                    "403": {
                        "description": "учётная запись заблокирована",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выход из системы (инвалидация refresh-токена)",
                "parameters": [
                    {
                        "description": "refresh_token (опционально)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление access-токена по refresh-токену",
                "parameters": [
                    {
                        "description": "refresh_token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.AuthRefreshResponse"
                        }
                    },
                    "401": {
                        "description": "недействительный refresh-токен",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    },
                    "403": {
                        "description": "учётная запись заблокирована",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "email + password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.AuthRegisterResponse"
                        }
                    },
                    "400": {
                        "description": "ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    },
                    "409": {
                        "description": "email уже существует",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Сменить пароль по токену из ссылки",
                "parameters": [
                    {
                        "description": "token + new_password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "недействительный/истёкший токен",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    },
                    "403": {
                        "description": "учётная запись заблокирована",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorPayload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_pro100kartochki_mozgoemka_internal_domain.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.AuthUserFull"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.AuthRefreshResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.AuthRegisterResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_pro100kartochki_mozgoemka_internal_domain.AuthUserBrief"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.AuthUserBrief": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.AuthUserFull": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_pro100kartochki_mozgoemka_internal_domain.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.ForgotPasswordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "reset_token": {
                    "type": "string"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "github_com_pro100kartochki_mozgoemka_internal_domain.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "github_com_pro100kartochki_mozgoemka_internal_domain.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "token"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}